+++ docs/changelog.rst
@@ -71,6 +71,9 @@ Added:
 * Support for the ``jp2`` file format using the imageformats plugin. To enable it, add
   ``imageformats = jp2`` to the ``[PLUGINS]`` section of your ``vimiv.conf``. Requires
   the qt imageformats plugin. Thanks `@szsdk`_ for testing!
+* The ``image.zoom_wheel_ctrl`` setting which toggles the need to hold the ``<ctrl>``
+  modifier for zooming an image with the mouse wheel. Thanks `@ArtemSmaznov`_ for the
+  idea!
 
 Changed:
 ^^^^^^^^
@@ -507,3 +510,4 @@ Initial release of the Qt version.
 .. _@Kakupakat: https://github.com/Kakupakat
 .. _@loiccoyle: https://github.com/loiccoyle
 .. _@szsdk: https://github.com/szsdk
+.. _@ArtemSmaznov: https://github.com/ArtemSmaznov